Removed code that was no longer being used. Moved the data parsing functions and their props to the Assets folder.
