+++ src/ClubInfo/ClubInfoPage.tsx
@@ -1,28 +1,14 @@
-import { Button, ButtonGroup } from '@chakra-ui/react';
-import React, { useState } from 'react';
+import React from 'react';
 import { useParams } from 'react-router-dom';
 
-import Vike_Labs_Icon from '../Assets/Images/Vike_Labs_Icon.png';
-import { ClubInfoData } from '../CustomProps';
-import data from '../JSON/ClubData.json';
+import { extractDate, findFullClubInfo } from '../Assets/DataParsing/Parsing';
 
 import { GridDiv, RightSubGridDiv, PhotoHeaderDiv } from './ClubInfoPageStyledComponents';
-import {
-  defaultLinks,
-  defaultDescription,
-  defaultTitle,
-  defaultPhotos,
-  defaultInfos,
-  // defaultLogo,
-  defaultTags,
-  defaultIcon,
-  defaultPath,
-} from './data';
+import { defaultPhotos /* defaultLogo,*/ } from './data';
 import Description from './Description';
-import Icon from './Icon';
 import InfoList from './Infos';
 import LinkList from './Links';
-import LogoItem, { Logo } from './Logo';
+import LogoItem from './Logo';
 import PathItem from './Path';
 import PhotoBar from './Photo';
 import TagList from './Tags';
@@ -58,7 +44,7 @@ function RightSubGrid(props: { subProps: subGridProps }) {
 function ClubInfoPage() {
   // Parsing and extracting the data from the JSON.
   const { clubType, clubName } = useParams<{ clubType: string; clubName: string }>();
-  const clubInfo = findClubInfo(clubName);
+  const clubInfo = findFullClubInfo(clubName);
   // Extracting the year and month the club was created.
   clubInfo[0].clubMembersAndCreationDate[1] = extractDate(clubInfo[0].clubMembersAndCreationDate[1]); // Commented this out until we have the acturate information.
   const membersAndDates = clubInfo[0].clubMembersAndCreationDate;
@@ -96,49 +82,6 @@ function ClubInfoPage() {
       </GridDiv>
     </div>
   );
-}
-
-function findClubInfo(clubName: string) {
-  let completedData: ClubInfoData[] = data
-    .filter((clubType) => clubType.name === clubName)
-    .map(({ name, description, tags, socialMedia, primaryEmail, members, createdDate }) => ({
-      clubName: name,
-      clubDescription: description,
-      clubTags: tags,
-      clubImage: Vike_Labs_Icon,
-      clubEmail: primaryEmail,
-      clubSocialMedia: socialMedia,
-      clubMembersAndCreationDate: [members + ' members', createdDate],
-    }));
-  return completedData;
-}
-
-function extractDate(rawDate: string) {
-  let extractedDate = RegExp(/[0-9]*\-[0-9]*\-[0-9]*/).exec(rawDate);
-  let processedDate = 'Date could not be extracted';
-  const months = [
-    'January',
-    'February',
-    'March',
-    'April',
-    'May',
-    'June',
-    'July',
-    'August',
-    'September',
-    'October',
-    'November',
-    'December',
-  ];
-
-  if (extractedDate != null) {
-    let splitDateArray = extractedDate[0].split('-');
-    let monthIndex = Number(splitDateArray[1]);
-    let extractedMonth = months[monthIndex - 1];
-    let extractedYear = splitDateArray[0];
-    processedDate = extractedMonth + ' ' + extractedYear;
-  }
-  return processedDate;
 }
 
 export default ClubInfoPage;
