+++ src/ClubInfo/ClubInfoPageStyledComponents.tsx
@@ -1,4 +1,3 @@
-import React, { useState } from 'react';
 import styled from 'styled-components';
 
 export const GridDiv = styled.div`
