+++ src/ClubInfo/Path.tsx
@@ -1,9 +1,6 @@
 import React from 'react';
 import { Link } from 'react-router-dom';
 import styled from 'styled-components';
-import { border, color, BorderProps, ColorProps } from 'styled-system';
-
-import { Path } from './data';
 
 const PathDiv = styled.p`
   @media screen and (max-width: 768px) {
