+++ src/ClubInfo/Title.tsx
@@ -1,8 +1,6 @@
 import React from 'react';
 import styled from 'styled-components';
 
-import { Title } from './data';
-
 const TitleDiv = styled.div`
   @media screen and (max-width: 768px) {
     margin-left: 0px;
