+++ src/JSON/ClubData.json
@@ -108,7 +108,7 @@
     "id": "8",
     "name": "UVic Students Offering Support (SOS)",
     "category": "Community Service & Volunteer",
-    "primaryEmail": "uvic. pres@studentsofferingsuppo rt.org",
+    "primaryEmail": "uvic.pres@studentsofferingsupport.org",
     "socialMedia": [
       "https://linktr.ee/UVicStudentsOfferingSupport"
     ],
@@ -499,7 +499,7 @@
     "id": "34",
     "name": "Ancient forest committee",
     "category": "Environmental",
-    "primaryEmail": "ancientforestuvic@gmail. com",
+    "primaryEmail": "ancientforestuvic@gmail.com",
     "socialMedia": [
       ""
     ],
@@ -544,7 +544,7 @@
     "id": "37",
     "name": "UVic Photography Excursion Club",
     "category": "Interest",
-    "primaryEmail": "photographyexcursionclub@ gmail.com",
+    "primaryEmail": "photographyexcursionclub@gmail.com",
     "socialMedia": [
       "https://www.facebook.com/uvicpec"
     ],
@@ -574,7 +574,7 @@
     "id": "39",
     "name": "Model United Nations",
     "category": "Interest",
-    "primaryEmail": "presidentuvicmun@gmail. com",
+    "primaryEmail": "presidentuvicmun@gmail.com",
     "socialMedia": [
       "https://www.facebook.com/UvicMUNClub"
     ],
@@ -679,7 +679,7 @@
     "id": "46",
     "name": "UVic Pre-Optometry Club",
     "category": "Interest",
-    "primaryEmail": "uvicpreoptometry@gmail. com",
+    "primaryEmail": "uvicpreoptometry@gmail.com",
     "socialMedia": [
       "https://www.instagram.com/uvicpreoptometry"
     ],
@@ -724,7 +724,7 @@
     "id": "49",
     "name": "Medieval Club",
     "category": "Interest",
-    "primaryEmail": "medievalclubuvic@gmail. com",
+    "primaryEmail": "medievalclubuvic@gmail.com",
     "socialMedia": [
       ""
     ],
@@ -739,7 +739,7 @@
     "id": "50",
     "name": "Model United Nations",
     "category": "Interest",
-    "primaryEmail": "presidentuvicmun@gmail. com",
+    "primaryEmail": "presidentuvicmun@gmail.com",
     "socialMedia": [
       "https://www.facebook.com/UvicMUNClub"
     ],
@@ -799,7 +799,7 @@
     "id": "54",
     "name": "Psyc After Undergrad",
     "category": "Interest",
-    "primaryEmail": "psycafterundergrad@gmail. com",
+    "primaryEmail": "psycafterundergrad@gmail.com",
     "socialMedia": [
       "https://www.instagram.com/psyc.after.undergrad/"
     ],
@@ -829,7 +829,7 @@
     "id": "56",
     "name": "Uvic NiteShifters Toastmasters Club",
     "category": "Interest",
-    "primaryEmail": "niteshiftersvictoriabc@gmail. com",
+    "primaryEmail": "niteshiftersvictoriabc@gmail.com",
     "socialMedia": [
       "https://www.facebook.com/uvictoastmasters"
     ],
@@ -844,7 +844,7 @@
     "id": "57",
     "name": "UVic Fashion Club",
     "category": "Interest",
-    "primaryEmail": "uvictoriafashionclub@gmail. com",
+    "primaryEmail": "uvictoriafashionclub@gmail.com",
     "socialMedia": [
       "https://www.instagram.com/uvicfashionclub/"
     ],
@@ -949,7 +949,7 @@
     "id": "64",
     "name": "UVic Financial Analysis Club (UVFAC)",
     "category": "Other",
-    "primaryEmail": "UVicFinancialAnalysisClub@ gmail.com",
+    "primaryEmail": "UVicFinancialAnalysisClub@gmail.com",
     "socialMedia": [
       ""
     ],
@@ -964,7 +964,7 @@
     "id": "65",
     "name": "Mental Health Awareness Club",
     "category": "Other",
-    "primaryEmail": "uvicmentalhealthclub@gmail .com",
+    "primaryEmail": "uvicmentalhealthclub@gmail.com",
     "socialMedia": [
       "http://www.facebook.com/uvicmentalhealthclub"
     ],
@@ -1009,7 +1009,7 @@
     "id": "68",
     "name": "Cabaret Voltaire Diversions",
     "category": "Other",
-    "primaryEmail": "cabaretvoltaireevents@gmail .com",
+    "primaryEmail": "cabaretvoltaireevents@gmail.com",
     "socialMedia": [
       "https://www.instagram.com/cabaretvoltairediversions/"
     ],
@@ -1204,7 +1204,7 @@
     "id": "81",
     "name": "Inter-Varsity Christian Fellowship Club",
     "category": "Religious",
-    "primaryEmail": "intervarsityuvictoria@gmail. com",
+    "primaryEmail": "intervarsityuvictoria@gmail.com",
     "socialMedia": [
       "https://www.instagram.com/uvic_ivcf"
     ],
@@ -1250,7 +1250,7 @@
     "id": "84",
     "name": "Inclusive Christians Club",
     "category": "Religious",
-    "primaryEmail": "courtneystokes506@yahoo. ca",
+    "primaryEmail": "courtneystokes506@yahoo.ca",
     "socialMedia": [
       ""
     ],
@@ -1310,7 +1310,7 @@
     "id": "88",
     "name": "UVic Poker Club",
     "category": "Sports & Recreation",
-    "primaryEmail": "dontraisemybigblind@gmail. com",
+    "primaryEmail": "dontraisemybigblind@gmail.com",
     "socialMedia": [
       ""
     ],
@@ -1431,7 +1431,7 @@
     "id": "96",
     "name": "UVic CSGO/VALORANT Club",
     "category": "Sports & Recreation",
-    "primaryEmail": "uvicvalorantcsgoclub@gmail. com",
+    "primaryEmail": "uvicvalorantcsgoclub@gmail.com",
     "socialMedia": [
       "https://www.facebook.com/groups/419162239460168"
     ],
@@ -1446,7 +1446,7 @@
     "id": "97",
     "name": "UVIC Skateboarding Club",
     "category": "Sports & Recreation",
-    "primaryEmail": "uvicskateboardingclub@gma il.com",
+    "primaryEmail": "uvicskateboardingclub@gmail.com",
     "socialMedia": [
       "https://www.instagram.com/uvicsc/"
     ],
@@ -1521,7 +1521,7 @@
     "id": "102",
     "name": "UVic Rocketyr",
     "category": "STEM",
-    "primaryEmail": "nicholascastellani@gmail. com",
+    "primaryEmail": "nicholascastellani@gmail.com",
     "socialMedia": [
       "https://www.facebook.com/uvicrocketry/"
     ],
@@ -1551,7 +1551,7 @@
     "id": "104",
     "name": "University of Victoria Renewable Energy Club",
     "category": "STEM",
-    "primaryEmail": "uvicrenewableenergyclub@g mail.com",
+    "primaryEmail": "uvicrenewableenergyclub@gmail.com",
     "socialMedia": [
       "https://www.facebook.com/University-of-Victoria-Renewable-Energy-Club-105152894858342"
     ],
@@ -1581,7 +1581,7 @@
     "id": "106",
     "name": "UVIC Pre-Dental Club",
     "category": "STEM",
-    "primaryEmail": "predentalclubuvic@gmail. com",
+    "primaryEmail": "predentalclubuvic@gmail.com",
     "socialMedia": [
       "https://www.instagram.com/predentalclubuvic/"
     ],
@@ -1656,7 +1656,7 @@
     "id": "111",
     "name": "UVic Game Dev",
     "category": "STEM",
-    "primaryEmail": "uvic.gamedev.club@gmail. com",
+    "primaryEmail": "uvic.gamedev.club@gmail.com",
     "socialMedia": [
       "https://www.facebook.com/groups/uvicgamedev"
     ],
@@ -1701,7 +1701,7 @@
     "id": "114",
     "name": "University of Victoria Formula Hybrid",
     "category": "STEM",
-    "primaryEmail": "uvicformulahybrid@gmail. com",
+    "primaryEmail": "uvicformulahybrid@gmail.com",
     "socialMedia": [
       "https://www.instagram.com/uvichybrid/"
     ],
@@ -1791,7 +1791,7 @@
     "id": "120",
     "name": "UVic Women in Science",
     "category": "STEM",
-    "primaryEmail": "womeninscienceuvic@gmail. com",
+    "primaryEmail": "womeninscienceuvic@gmail.com",
     "socialMedia": [
       "https://www.facebook.com/womeninscienceuvic"
     ],
@@ -1806,7 +1806,7 @@
     "id": "121",
     "name": "Students Above Stigma",
     "category": "STEM",
-    "primaryEmail": "uvic.students.above. stigma@gmail.com",
+    "primaryEmail": "uvic.students.above.stigma@gmail.com",
     "socialMedia": [
       "https://www.instagram.com/uvic_sas/"
     ],
@@ -1821,7 +1821,7 @@
     "id": "122",
     "name": "UVic Competitive Programming Club",
     "category": "STEM",
-    "primaryEmail": "uvic.programming. club@gmail.com",
+    "primaryEmail": "uvic.programming.club@gmail.com",
     "socialMedia": [
       ""
     ],
